src/main/java/com/mo/Service/Impl/ProductCategoryServiceImpl.java
package com.mo.Service.Impl;

import com.mo.Entity.ProductCategory;
import com.mo.Repository.ProductCategoryRepository;
import com.mo.Service.ProductCategoryService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

/**
 * @author 音神
 * @date 2018/10/19 15:15
 */

@Service
public class ProductCategoryServiceImpl implements ProductCategoryService {

    @Autowired
    private ProductCategoryRepository repository;

    @Override
    public ProductCategory findOne(Integer categoryId) {
        return repository.findById(categoryId).get();
    }

    @Override
    public List<ProductCategory> findAll() {
        return repository.findAll();
    }

    @Override
    public List<ProductCategory> findByCategoryTypeIn(List<Integer> ProductCategoryTypeList) {
        return repository.findByCategoryTypeIn(ProductCategoryTypeList);
    }

    @Override
    public ProductCategory save(ProductCategory productCategory) {
        return repository.save(productCategory);
    }
}
